refactor: Fix files and variables naming
